prevent non-root

+++ install/image.sh
@@ -1,5 +1,10 @@
 #! /usr/bin/env bash
 
+if [ $EUID -ne 0 ]; then
+    echo "This script must be run as root"
+    exit 1
+fi
+
 set -exuo pipefail
 
 rm disk.raw
